refactor: reorganize tailwind package so we dont import server code on client
