+++ packages/ui/tailwind/src/config/createVarsForColors/createVarsForColors.spec.js
