+++ packages/ui/tailwind/src/config/loadThemesConfig.js
@@ -1,10 +1,10 @@
 import defaultsDeep from 'lodash/defaultsDeep.js';
 import { getEvidenceConfig } from '@evidence-dev/sdk/config';
 
-import { ThemesConfigFileSchema } from './schemas/config.js';
-import { defaultThemesConfig } from './defaultThemesConfig.js';
+import { ThemesConfigFileSchema } from '../schemas/config.js';
+import { defaultThemesConfig } from '../themes/defaultThemesConfig.js';
 
-/** @returns {import('./schemas/types.js').ThemesConfig} */
+/** @returns {import('../schemas/types.js').ThemesConfig} */
 export const loadThemesConfig = () => {
 	let fromEvidenceConfig = {};
 	try {
