+++ packages/ui/tailwind/src/schemas/colorPalettes.js
