+++ packages/ui/tailwind/src/schemas/colorScales.js
