+++ packages/ui/tailwind/src/schemas/colors.js
