+++ packages/ui/tailwind/src/themes/index.js
@@ -1,4 +1,2 @@
-export * from './vite-plugin/index.js';
-export * from './loadThemesConfig.js';
 export * from './buildThemes.js';
-export * from './schemas/type-guards.js';
+export * from '../schemas/type-guards.js';
