--- packages/ui/tailwind/src/themes/schemas/index.js
@@ -1,5 +1,0 @@
-export * from './colorPalettes.js';
-export * from './colors.js';
-export * from './colorScales.js';
-export * from './config.js';
-export * from './type-guards.js';
