--- packages/ui/tailwind/src/themes/types.d.ts
@@ -1,2 +1,0 @@
-export * from './schemas/types.d.ts';
-export * from './vite-plugin/evidence-themes.d.ts';
