--- packages/ui/tailwind/src/themes/vite-plugin/index.js
@@ -1,1 +1,0 @@
-export * from './vite-plugin.js';
