+++ packages/ui/tailwind/src/types.d.ts
@@ -1,2 +1,3 @@
-export * from './themes/types.js';
+export * from './vite-plugin/evidence-themes.d.ts';
+export * from './schemas/types.d.ts';
 export * from './index.js';
